+++ components/shapediver/viewport/ViewportIcons.tsx
@@ -4,18 +4,26 @@ import {useViewportId} from "@AppBuilderShared/hooks/shapediver/viewer/useViewpo
 import {useShapeDiverStoreParameters} from "@AppBuilderShared/store/useShapeDiverStoreParameters";
 import {useShapeDiverStoreViewport} from "@AppBuilderShared/store/useShapeDiverStoreViewport";
 import {useShapeDiverViewportIconsStore} from "@AppBuilderShared/store/useShapeDiverViewportIconsStore";
-import {
-	ButtonRenderContext,
-	renderViewportIcons,
-} from "@AppBuilderShared/types/components/shapediver/componentTypes";
+import {ButtonRenderContext} from "@AppBuilderShared/types/components/shapediver/componentTypes";
 import {ViewportTransparentBackgroundStyle} from "@AppBuilderShared/types/shapediver/viewport";
 import {
 	ViewportIconsOptionalProps,
 	ViewportIconsProps,
 } from "@AppBuilderShared/types/shapediver/viewportIcons";
-import {Paper, Transition, useProps} from "@mantine/core";
+import {
+	Divider,
+	DividerProps,
+	Paper,
+	Transition,
+	useProps,
+} from "@mantine/core";
 import React, {useCallback, useMemo, useState} from "react";
 import {useShallow} from "zustand/react/shallow";
+import {ViewportTypeToIcon} from "~/shared/context/ComponentContext";
+import {
+	ViewportIconButtonEnum,
+	ViewportIconLayoutItem,
+} from "~/shared/types/store/shapediverStoreViewportIcons";
 import ViewportOverlayWrapper from "./ViewportOverlayWrapper";
 
 export const defaultStyleProps: ViewportIconsOptionalProps = {
@@ -47,6 +55,162 @@ export const defaultStyleProps: ViewportIconsOptionalProps = {
 		timingFunction: "ease",
 		keepMounted: true,
 	},
+};
+
+const renderButtonByKind = (kind: ViewportIconButtonEnum, context: any) => {
+	const {
+		viewport,
+		namespace,
+		buttonsDisabled,
+		executing,
+		hasPendingChanges,
+		iconsVisible,
+		fullscreenId,
+		...commonProps
+	} = context;
+
+	switch (kind) {
+		case ViewportIconButtonEnum.Ar:
+			return React.createElement(
+				ViewportTypeToIcon[ViewportIconButtonEnum.Ar],
+				{key: "ar", viewport, ...commonProps},
+			);
+		case ViewportIconButtonEnum.Zoom:
+			return React.createElement(
+				ViewportTypeToIcon[ViewportIconButtonEnum.Zoom],
+				{key: "zoom", viewport, ...commonProps},
+			);
+		case ViewportIconButtonEnum.Fullscreen:
+			return React.createElement(
+				ViewportTypeToIcon[ViewportIconButtonEnum.Fullscreen],
+				{
+					key: "fullscreen",
+					fullscreenId,
+					enableFullscreenBtn: true,
+					...commonProps,
+				},
+			);
+		case ViewportIconButtonEnum.Cameras:
+			return React.createElement(
+				ViewportTypeToIcon[ViewportIconButtonEnum.Cameras],
+				{
+					key: "cameras",
+					viewport,
+					visible: iconsVisible,
+					...commonProps,
+				},
+			);
+		case ViewportIconButtonEnum.Undo:
+			return React.createElement(
+				ViewportTypeToIcon[ViewportIconButtonEnum.Undo],
+				{
+					key: "undo",
+					disabled: buttonsDisabled || executing || hasPendingChanges,
+					hasPendingChanges,
+					executing,
+					...commonProps,
+				},
+			);
+		case ViewportIconButtonEnum.Redo:
+			return React.createElement(
+				ViewportTypeToIcon[ViewportIconButtonEnum.Redo],
+				{
+					key: "redo",
+					disabled: buttonsDisabled || executing || hasPendingChanges,
+					hasPendingChanges,
+					executing,
+					...commonProps,
+				},
+			);
+		case ViewportIconButtonEnum.Reload:
+			return React.createElement(
+				ViewportTypeToIcon[ViewportIconButtonEnum.Reload],
+				{
+					key: "reload",
+					disabled:
+						!namespace ||
+						buttonsDisabled ||
+						executing ||
+						hasPendingChanges,
+					namespace: namespace || "",
+					hasPendingChanges,
+					executing,
+					...commonProps,
+				},
+			);
+		case ViewportIconButtonEnum.HistoryMenu:
+			return React.createElement(
+				ViewportTypeToIcon[ViewportIconButtonEnum.HistoryMenu],
+				{
+					key: "historyMenu",
+					disabled:
+						!namespace || buttonsDisabled || hasPendingChanges,
+					namespace: namespace || "",
+					visible: iconsVisible,
+					...commonProps,
+				},
+			);
+		default:
+			return null;
+	}
+};
+
+const renderViewportIcons = (
+	viewportIcons: ViewportIconLayoutItem[],
+	buttonContext: ButtonRenderContext,
+	dividerProps: DividerProps = {},
+) => {
+	const sections: React.ReactNode[] = [];
+	if (viewportIcons.length === 0) return sections;
+
+	viewportIcons.forEach((item, index) => {
+		if (item.type === "button") {
+			const button = renderButtonByKind(item.button.type, buttonContext);
+			if (button) sections.push(button);
+		} else if (item.type === "group") {
+			const groupButtons: React.ReactNode[] = [];
+			item.sections.forEach((section) => {
+				section.forEach((buttonDef) => {
+					const button = renderButtonByKind(
+						buttonDef.type,
+						buttonContext,
+					);
+					if (button) groupButtons.push(button);
+				});
+				// Add divider between sections within a group
+				if (
+					groupButtons.length > 0 &&
+					section !== item.sections[item.sections.length - 1]
+				) {
+					groupButtons.push(
+						React.createElement(Divider, {
+							key: `divider-${index}-${section.length}`,
+							...dividerProps,
+						}),
+					);
+				}
+			});
+			sections.push(
+				React.createElement(
+					React.Fragment,
+					{key: `group-${index}`},
+					...groupButtons,
+				),
+			);
+		}
+
+		// Add divider between layout items
+		if (index < viewportIcons.length - 1) {
+			sections.push(
+				React.createElement(Divider, {
+					key: `layout-divider-${index}`,
+					...dividerProps,
+				}),
+			);
+		}
+	});
+
+	return sections;
 };
 
 export default function ViewportIcons(
